refactor(core)!: rename liveIconCollections to defineLiveIconCollections
Puts both config-file builders in the define* family (Astro's own
defineConfig/defineCollection/defineLiveCollection convention):
defineIconCollection(sources) defines one build collection,
defineLiveIconCollections({...}) defines a keyed set of live ones - the
plural and object form stay, since writing the key once is the live-side
feature. Module and test files renamed to match.
