+++ .changeset/loud-icons-refactor.md
@@ -15,11 +15,11 @@ Rebuilt astro-icon on top of Astro's Content Layer instead of a custom Vite reso
   ```
   `defineIconCollection(sources)` is sugar for `defineCollection({ loader: createIconLoader(sources) })` - drop down to that form (both still exported) to attach your own Zod `schema`.
   `<Icon name="mdi:search" />` resolves the `mdi` collection's `search` entry. `<Icon name="search" />` (no prefix) resolves a collection literally named `icons`, a naming convention rather than something astro-icon enforces, so you must define a collection called `icons` yourself for bare names to work.
-- **The package has five purpose-named entries.** `astro-icon/components` (`Icon`, `LiveIcon`, plus exported `IconProps`/`LiveIconProps` for typing wrapper components), `astro-icon/collections` (everything a `content.config.ts`/`live.config.ts` needs: `defineIconCollection`, `liveIconCollections`, the `iconify`/`iconifyApi`/`localSvg` sources, and the `createIconLoader`/`createLiveIconLoader` escape hatches), `astro-icon/source` (the custom-source authoring kit: `defineIconSource`, `entryFromSVG`, `entryFromIconifyData`, `mergeSources`), `astro-icon/optimize` (the opt-in SVGO wrapper), and the root `astro-icon` (every public type plus `AstroIconError` - safe to import from any context).
+- **The package has five purpose-named entries.** `astro-icon/components` (`Icon`, `LiveIcon`, plus exported `IconProps`/`LiveIconProps` for typing wrapper components), `astro-icon/collections` (everything a `content.config.ts`/`live.config.ts` needs: `defineIconCollection`, `defineLiveIconCollections`, the `iconify`/`iconifyApi`/`localSvg` sources, and the `createIconLoader`/`createLiveIconLoader` escape hatches), `astro-icon/source` (the custom-source authoring kit: `defineIconSource`, `entryFromSVG`, `entryFromIconifyData`, `mergeSources`), `astro-icon/optimize` (the opt-in SVGO wrapper), and the root `astro-icon` (every public type plus `AstroIconError` - safe to import from any context).
 - **`iconifyApi` can target a self-hosted Iconify API.** Pass `host: "https://icons.example.com"` to resolve icons from [your own API instance](https://iconify.design/docs/api/hosting.html) instead of the public `api.iconify.design`.
 - **A collection is always exactly what its source(s) allow, restrict it with `allowed: [...]`.** `iconify("mdi")` with no options loads (and types) the entire `mdi` pack, provided it's installed locally. `iconifyApi("mdi", { allowed: [...] })` resolves from the public Iconify API instead, which can only ever return specific icons you ask for, never "everything", so `allowed` is required there. Pass both as an array (`defineIconCollection([iconify(...), iconifyApi(...)])`) for "prefer local, fall back to the API." Pass `allowed: ["account", "home"]` to restrict a pack to those two, a deliberate allowlist, not a scan-based guess, so loading and generated types are always the same set.
 - **Everything is built on one primitive: `IconSource`** (`{ name, getIcons(names), listIcons?() }`). `getIcons` always takes an array and returns a `Map<string, IconEntry | Error>` covering every name asked for - even a single-icon lookup is a batch of one - so a source backed by a real batch endpoint (`iconifyApi`'s `?icons=a,b,c`) only pays for one request no matter how many names a caller asks for in one call, and a source with nothing to batch (`localSvg`, an already-loaded Iconify pack) just resolves each name independently. `iconify(pack, options)`/`iconifyApi(pack, options)` are the Iconify-backed ones, fed into `defineIconCollection()`. Write your own `IconSource` for a non-Iconify pack, a custom npm package's icons, or any other API.
-- **Combine multiple sources into one collection** by passing an array - every collection-building call (`defineIconCollection`, `liveIconCollections`' values, the loader factories) accepts `IconSource | IconSource[]`. Each requested icon is resolved by trying sources in order, first match wins:
+- **Combine multiple sources into one collection** by passing an array - every collection-building call (`defineIconCollection`, `defineLiveIconCollections`' values, the loader factories) accepts `IconSource | IconSource[]`. Each requested icon is resolved by trying sources in order, first match wins:
   ```ts
   import { defineIconCollection, iconify } from "astro-icon/collections";
 
@@ -31,8 +31,8 @@ Rebuilt astro-icon on top of Astro's Content Layer instead of a custom Vite reso
   };
   ```
   `mergeSources` (from `astro-icon/source`) pre-composes an array into a single `IconSource` object for the cases a config-file array can't express, e.g. a library exporting one composed source.
-- **New `<LiveIcon>` component**, backed by Astro's experimental live content collections. `astro-icon/collections` exports `liveIconCollections({ key: source })` plus `iconify`/`iconifyApi`, so a live source isn't limited to Iconify: implement your own `IconSource` to back `<LiveIcon>` with a different pack format or API. Unlike `<Icon>`, `<LiveIcon>` always requires a `collection:name` value; there's no default collection.
-  - `liveIconCollections` spreads into `live.config.ts`'s `collections` object; each key becomes both the collection's registration key and its generated `LiveCollectionName` type, written exactly once. To call Astro's `defineLiveCollection()` yourself, use `createLiveIconLoader(source, { collection })` and pass the key explicitly - Astro only reveals the real key at request time, so the loader warns (and corrects typegen) on the first request if the two don't match.
+- **New `<LiveIcon>` component**, backed by Astro's experimental live content collections. `astro-icon/collections` exports `defineLiveIconCollections({ key: source })` plus `iconify`/`iconifyApi`, so a live source isn't limited to Iconify: implement your own `IconSource` to back `<LiveIcon>` with a different pack format or API. Unlike `<Icon>`, `<LiveIcon>` always requires a `collection:name` value; there's no default collection.
+  - `defineLiveIconCollections` spreads into `live.config.ts`'s `collections` object; each key becomes both the collection's registration key and its generated `LiveCollectionName` type, written exactly once. To call Astro's `defineLiveCollection()` yourself, use `createLiveIconLoader(source, { collection })` and pass the key explicitly - Astro only reveals the real key at request time, so the loader warns (and corrects typegen) on the first request if the two don't match.
   - `getLiveCollection(collection, { ids: [...] })` resolves that specific list of names in one batched call, instead of one `<LiveIcon>`/`getLiveEntry()` per name - reach for it when every name is already known before any of them are fetched (a search result set, saved user picks), rather than resolving each independently. It also warms `<LiveIcon>`'s per-entry cache, so a later individual lookup for one of the same names is a cache hit.
 - **Local icons are back, as a proper `IconSource`.** `localSvg(dir?, options?)` (from `astro-icon/collections`) loads every `.svg` under a directory (default `src/icons/`). Subdirectories become part of the name (`src/icons/logos/deno.svg` is `"logos/deno"`), and it watches the directory in dev, the same way Astro's built-in file-backed collections (e.g. markdown) do: add, edit, or remove a file and the collection picks it up with no dev server restart. This is the suggested default for the `icons` collection:
   ```ts
