+++ examples/platform-builder/src/live.config.ts
@@ -1,4 +1,4 @@
-import { iconifyApi, liveIconCollections } from "astro-icon/collections";
+import { iconifyApi, defineLiveIconCollections } from "astro-icon/collections";
 import { brandKitSource } from "./lib/brandKitSource";
 
 // Live collections resolve per request. Reach for one when the icon names
@@ -6,14 +6,14 @@ import { brandKitSource } from "./lib/brandKitSource";
 // `nav` collection in content.config.ts) but genuinely unknowable at build
 // time, because they depend on what a user does or what another system holds.
 //
-// `liveIconCollections()` writes each key once - it becomes both the Astro
+// `defineLiveIconCollections()` writes each key once - it becomes both the Astro
 // collection key and the generated `LiveCollectionName` type, so there's no
 // separate `name`/`collection` option to keep in sync by hand.
 export const collections = {
   // No `allowed` allowlist and nothing installed: each requested name is fetched
   // individually from api.iconify.design. A build-time collection can't express
   // this - it would have to enumerate every candidate up front.
-  ...liveIconCollections({
+  ...defineLiveIconCollections({
     ph: iconifyApi("ph"),
 
     // A custom IconSource.
