+++ packages/core/components/LiveIcon.astro
@@ -14,7 +14,7 @@ import type { IconEntry } from "../typings/types";
  * names ahead of time, such as a user-driven icon search.
  *
  * Requires a live collection registered in `live.config.ts` (see
- * `liveIconCollections()`); live content
+ * `defineLiveIconCollections()`); live content
  * collections are stable as of Astro 6, no experimental flag needed.
  * If the icon fails to resolve,
  * `<LiveIcon>` logs a warning and renders nothing rather than throwing, so
