+++ packages/core/src/collections.ts
@@ -1,15 +1,15 @@
 /**
  * Everything that goes in a config file - `content.config.ts` or `live.config.ts`.
  *
- * `defineIconCollection`/`liveIconCollections` build the collections; `iconify`/`iconifyApi`/
+ * `defineIconCollection`/`defineLiveIconCollections` build the collections; `iconify`/`iconifyApi`/
  * `localSvg` are the bundled sources that feed them; `createIconLoader`/`createLiveIconLoader`
  * are the drop-down-a-layer escape hatches for `defineCollection({ loader, schema })`/
  * `defineLiveCollection()`. Authoring a *custom* source lives in `astro-icon/source` instead.
  */
 export { defineIconCollection } from "./content/defineIconCollection.js";
 export type { IconCollectionConfig } from "./content/defineIconCollection.js";
-export { liveIconCollections } from "./content/liveCollections.js";
-export type { LiveIconCollectionConfig } from "./content/liveCollections.js";
+export { defineLiveIconCollections } from "./content/defineLiveIconCollections.js";
+export type { LiveIconCollectionConfig } from "./content/defineLiveIconCollections.js";
 export { createIconLoader } from "./content/loader.js";
 export type { IconLoaderOptions } from "./content/loader.js";
 export { createLiveIconLoader } from "./content/liveLoader.js";
