+++ packages/core/test/fixtures/basic/src/live.config.ts
@@ -4,7 +4,7 @@ import {
   iconify,
 } from "astro-icon/collections";
 
-// The raw registration form; the `live` fixture covers `liveIconCollections()`.
+// The raw registration form; the `live` fixture covers `defineLiveIconCollections()`.
 export const collections = {
   liveSpinners: defineLiveCollection({
     loader: createLiveIconLoader(iconify("svg-spinners"), {
